fix: 修复 Qt::TextWordBreak 标识符未声明编译错误
- 将 Qt::TextWordBreak 与 Qt::AlignCenter 显式转换为 int 后进行位或运算，以适配 QPainter::drawText 的参数要求。
- 确保在严苛模式下标识符解析正确。

+++ src/ui/ContentPanel.cpp
@@ -1049,8 +1049,8 @@ void GridItemDelegate::paint(QPainter* painter, const QStyleOptionViewItem& opti
     textFont.setPointSize(8);
     textFont.setBold(false);
     painter->setFont(textFont);
-    // 2026-03-xx 按照用户要求：开启名称自动换行渲染
-    painter->drawText(nameRect.adjusted(4, 0, -4, 0), Qt::AlignCenter | Qt::TextWordBreak, name);
+    // 2026-03-xx 按照用户要求：开启名称自动换行渲染 (修复编译标识符未声明错误)
+    painter->drawText(nameRect.adjusted(4, 0, -4, 0), static_cast<int>(Qt::AlignCenter) | static_cast<int>(Qt::TextWordBreak), name);
 
     painter->restore();
 }
